code added

+++ fullStackApp/backend/routes/postRoute.js
@@ -8,8 +8,31 @@ const postRouter = express.Router();
 
 postRouter.get('/', auth, async(req, res) => {
     const userID = req.body.userID;
+    const minComments = req.query.minComments;
+    const maxComments = req.query.maxComments;
     try {
-        const posts = await PostModel.find({userID});
+        let query = {userID};
+        if(minComments){
+            query.no_of_comments = {$gte : +minComments}
+        }
+
+        if(maxComments){
+            query.no_of_comments = {...query.no_of_comments, $lte : +maxComments}
+        }
+
+        const posts = await PostModel.find(query);
+        res.status(200).send(posts);
+    } catch (error) {
+        res.status(400).send({'msg' : error.message});
+    }
+})
+
+postRouter.get('/top', auth, async(req, res) => {
+    const id = req.body.userID.toString();
+    const page = +req.query.page || 1;
+    const limit = 3;
+    try {
+        const posts = await PostModel.find({userID : id}).sort({['no_of_comments'] : -1}).skip((page - 1)*limit).limit(limit);
         res.status(200).send(posts);
     } catch (error) {
         res.status(400).send({'msg' : error.message});
@@ -35,9 +58,10 @@ postRouter.post('/add', auth, async(req, res) => {
 
 postRouter.patch('/update/:id', auth, async(req, res) => {
     const id = req.params.id;
-    const user = await PostModel.findOne({_id : id});
+    const post = await PostModel.findOne({_id : id});
+
     try {
-        if(user.userID.toString() == req.body.userID){
+        if(post.userID.toString() == req.body.userID){
             const updatedPost = await PostModel.findByIdAndUpdate({_id : id}, req.body, {new : true});
             res.status(200).send({'msg' : 'Post updated successfully', updatedPost});
         }
@@ -52,9 +76,16 @@ postRouter.patch('/update/:id', auth, async(req, res) => {
 
 postRouter.delete('/delete/:id', auth, async(req, res) => {
     const id = req.params.id;
+    const post = await PostModel.findOne({_id : id});
+
     try {
-        const deletedPost = await PostModel.findByIdAndDelete({_id : id});
-        res.status(200).send({'msg' : 'Post deleted successfully'});
+        if(post.userID.toString() == req.body.userID){
+            const deletedPost = await PostModel.findByIdAndDelete({_id : id});
+            res.status(200).send({'msg' : 'Post deleted successfully'});
+        }
+        else{
+            res.status(400).send({'msg' : 'You are not authorized to delete the post'})
+        }
     } catch (error) {
         res.status(400).send({'msg' : error.message});
     }
